replaced review, user and contacts handler of api
